+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the extension will be documented in this file.
 
+### 0.0.2 - 2021-11-11
+
+1. use ! to announce word boundary.
+2. apply settings without reloading when changed.
+
 ### 0.0.1 - 2021-11-10
 
 - Initial release